refactored business rules to fluentvalidation, and refactored some services to private functions

+++ src/Application/Features/UserFavouriteProducts/Commands/Delete/DeleteUserFavouriteProductCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿using Core.Application.GenericRepository;
+﻿using Application.Features.UserFavouriteProducts.Constants;
+using Core.Application.GenericRepository;
+using Core.CrossCuttingConcerns.Exceptions.Types;
 using Domain.Models;
 using MediatR;
 
@@ -16,6 +18,8 @@ namespace Application.Features.UserFavouriteProducts.Commands.Delete
         public async Task<DeletedUserFavouriteProductResponse> Handle(DeleteUserFavouriteProductCommand request, CancellationToken cancellationToken)
         {
             var userFavouriteProductToDelete = await _userFavouriteProductRepository.GetAsync(predicate: p => (p.ProductId == request.ProductId) && (p.UserId == request.UserId));
+
+            if (userFavouriteProductToDelete == null) throw new BusinessException(UserFavouriteProductBusinessMessages.UserFavouriteProductMustExistToDelete);
 
             var deletedUserFavouriteProduct = await _userFavouriteProductRepository.DeleteAsync(userFavouriteProductToDelete, permanent: true);
